# specs/001-zero-training-mvp/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: LiverRa v1 MVP HTTP API
  version: 1.0.0
  description: |
    Public HTTP API for the LiverRa zero-training liver AI pipeline platform.

    **Auth**: OIDC Bearer token from AWS Cognito (research A.1) with `custom:tenant_id` claim.
    **RBAC**: Every protected op declares `x-required-permission`. Step-up auth (≤5 min freshness) is marked `x-step-up: true`.
    **Errors**: RFC 7807 Problem Details. `type`, `title`, `status`, `detail`, plus LiverRa extensions `tenant_id`, `audit_event_id`, `claim_key` (never PHI).
    **Audit**: Every state-changing op emits a FHIR AuditEvent (chain-of-hashes research A.3).
servers:
  - url: https://app.liverra.ai/api/v1
    description: Production
  - url: http://localhost:7050/api/v1
    description: Local dev (FastAPI ml-inference)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuditEventSummary:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        actor: { type: string }
        subject: { type: string }
        timestamp: { type: string, format: date-time }
        outcome: { type: string, enum: [success, denied, error] }
        chain_sequence_no: { type: integer }
...
paths:
...
  /admin/users/invite:
    post:
      tags: [admin]
      x-required-permission: admin.invite_user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [hpb_surgeon, radiologist, fellow, ops, compliance, dpo] }
                display_name: { type: string }
                locale_preference: { type: string, enum: [en, de, ka] }
      responses:
        '202': { description: Invite email queued (72h token) }
  /admin/users/{user_id}/suspend:
    parameters:
      - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [admin]
      x-required-permission: admin.suspend_user
      responses:
        '204': { description: Suspended (historical records preserved) }
  /admin/pacs-destination:
    put:
      tags: [admin]
      x-required-permission: admin.configure_pacs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ae_title: { type: string }
                host: { type: string }
                port: { type: integer }
                use_tls: { type: boolean }
                cert_fingerprint: { type: string, nullable: true }
      responses:
        '200':
          description: Saved (after successful C-ECHO pre-flight)
          content:
            application/json:
              schema:
                type: object
                properties:
                  destination: { $ref: '#/components/schemas/Tenant/properties/pacs_destination' }
                  cecho_round_trip_ms: { type: integer }
...
  /admin/studies/{study_id}/delete-request:
    parameters:
      - { name: study_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [admin]
      x-required-permission: admin.approve_deletion
      summary: Approve a user-submitted case-deletion request (FR-046)
      responses:
        '200': { description: Approved + soft-deleted }
  /admin/audit:
    get:
      tags: [admin]
      x-required-permission: admin.view_audit
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEventSummary' }
...